Checkout repo before setting up go
Version 4 of the setup-go action has caching enabled by default,
but reads the existing go.mod and go.sum files in the checkout.
If these do not exist, it issues a warning and avoids the cache
behavior. To prevent this, this commit follows the suggested
practice of performing the code checkout before running the
setup-go action.