Add cluster ping to publish
If the IP address for a cluster is not reachable, it takes
a very long time for the mount to time out, holding up K8S
operations and preventing quick deletion of incorrect volumes.

This change adds ping operation to ensure that the node is
able to communicate with the expected cluster before attempting
the mount operation. The result is cached for 20 seconds to
prevent overwhelming the API if many volumes are being
published in a short period of time.